yappy-modules/semantic-chunker/requirements.txt
# gRPC dependencies
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3

# NLP dependencies for semantic chunking
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1
nltk==3.8.1

# Testing dependencies
pytest==7.4.4
pytest-cov==4.1.0

# Utility dependencies
python-dotenv==1.0.1
pyyaml==6.0.1